refactor: split large formatter files into smaller modules
Refactor formatters.py (38KB) and test_analytics_formatters.py (38KB)
to comply with 20KB file size limit enforced by pre-commit hooks.

Changes:
- Split src/nhl_scrabble/analytics/formatters.py into 11 modules:
  - formatters/__init__.py: Re-exports all formatter classes
  - formatters/text_formatter.py: TextFormatter (5.9KB)
  - formatters/json_formatter.py: JSONFormatter (2.4KB)
  - formatters/html_formatter.py: HTMLFormatter (6.4KB)
  - formatters/yaml_formatter.py: YAMLFormatter (2.9KB)
  - formatters/xml_formatter.py: XMLFormatter (3.7KB)
  - formatters/table_formatter.py: TableFormatter (2.5KB)
  - formatters/markdown_formatter.py: MarkdownFormatter (3.4KB)
  - formatters/csv_formatter.py: CSVFormatter (3.2KB)
  - formatters/excel_formatter.py: ExcelFormatter (4.7KB)
  - formatters/template_formatter.py: TemplateFormatter (4.1KB)

- Split tests/unit/test_analytics_formatters.py into 10 test modules:
  - test_text_formatter.py (7.2KB)
  - test_json_formatter.py (5.4KB)
  - test_html_formatter.py (7.1KB)
  - test_yaml_formatter.py (2.7KB)
  - test_xml_formatter.py (3.1KB)
  - test_table_formatter.py (2.3KB)
  - test_markdown_formatter.py (3.0KB)
  - test_csv_formatter.py (2.5KB)
  - test_excel_formatter.py (3.4KB)
  - test_template_formatter.py (3.4KB)

- Remove exclusions from scripts/check_file_size.sh
  (reverted previous incorrect approach of ignoring large files)

All 54 formatter tests passing. All files now under 20KB limit.

Addresses user feedback: refactor large files instead of adding
them to ignore list in check_file_size.sh.
